Split missing factory into its own file
